+++ app/xml_cdr/xml_cdr.php
@@ -183,12 +183,19 @@
 	if ($_GET['call_result'] != 'missed') {
 		echo button::create(['type'=>'button','label'=>$text['button-missed'],'icon'=>'phone-slash','link'=>'?call_result=missed']);
 	}
-	echo button::create(['type'=>'button','label'=>$text['button-export'],'icon'=>$_SESSION['theme']['button_icon_export'],'onclick'=>"toggle_select('export_format'); this.blur();"]);
-	echo 		"<select class='formfld' style='display: none; width: auto;' name='export_format' id='export_format' onchange=\"display_message('".$text['message-preparing_download']."'); toggle_select('export_format'); document.getElementById('frm_export').submit();\">";
-	echo "			<option value='' disabled='disabled' selected='selected'>".$text['label-format']."</option>";
-	echo "			<option value='csv'>CSV</option>";
-	echo "			<option value='pdf'>PDF</option>";
-	echo "		</select>";
+
+	if (permission_exists('xml_cdr_export')) {
+		echo button::create(['type'=>'button','label'=>$text['button-export'],'icon'=>$_SESSION['theme']['button_icon_export'],'onclick'=>"toggle_select('export_format'); this.blur();"]);
+		echo 		"<select class='formfld' style='display: none; width: auto;' name='export_format' id='export_format' onchange=\"display_message('".$text['message-preparing_download']."'); toggle_select('export_format'); document.getElementById('frm_export').submit();\">";
+		echo "			<option value='' disabled='disabled' selected='selected'>".$text['label-format']."</option>";
+		if (permission_exists('xml_cdr_export_csv')) {
+			echo "			<option value='csv'>CSV</option>";
+		}
+		if (permission_exists('xml_cdr_export_pdf')) {
+			echo "			<option value='pdf'>PDF</option>";
+		}
+		echo "		</select>";
+	}
 	if (!$archive_request && permission_exists('xml_cdr_delete')) {
 		echo button::create(['type'=>'button','label'=>$text['button-delete'],'icon'=>$_SESSION['theme']['button_icon_delete'],'name'=>'btn_delete','onclick'=>"modal_open('modal-delete','btn_delete');"]);
 	}
@@ -622,17 +629,6 @@
 		//loop through the results
 			$x = 0;
 			foreach ($result as $index => $row) {
-				//get the date and time
-					$tmp_year = date("Y", strtotime($row['start_stamp']));
-					$tmp_month = date("M", strtotime($row['start_stamp']));
-					$tmp_day = date("d", strtotime($row['start_stamp']));
-					$tmp_start_epoch_date = escape(date("j M Y", $row['start_epoch']));
-					if ($_SESSION['domain']['time_format']['text'] == '12h') {
-						$tmp_start_epoch_time = escape(date("g:i:s a", $row['start_epoch']));
-					}
-					else {
-						$tmp_start_epoch_time = escape(date("H:i:s", $row['start_epoch']));
-					}
 
 				//get the hangup cause
 					$hangup_cause = $row['hangup_cause'];
@@ -755,7 +751,7 @@
 					}
 				//recording
 					if (permission_exists('xml_cdr_recording') && (permission_exists('xml_cdr_recording_play') || permission_exists('xml_cdr_recording_download'))) {
-						if ($record_path != '' && file_exists($record_path.'/'.$record_name)) {
+						if ($record_path != '') {
 							$content .= "	<td class='middle button center no-link no-wrap'>";
 							if (permission_exists('xml_cdr_recording_play')) {
 								$content .= 	"<audio id='recording_audio_".escape($row['xml_cdr_uuid'])."' style='display: none;' preload='none' ontimeupdate=\"update_progress('".escape($row['xml_cdr_uuid'])."')\" onended=\"recording_reset('".escape($row['xml_cdr_uuid'])."');\" src=\"download.php?id=".escape($row['xml_cdr_uuid'])."&t=record\" type='".escape($record_type)."'></audio>";
@@ -784,12 +780,12 @@
 					}
 				//start
 					if (permission_exists('xml_cdr_start')) {
-						$content .= "	<td class='middle right no-wrap'>".$tmp_start_epoch_date."</td>\n";
-						$content .= "	<td class='middle right no-wrap hide-md-dn'>".$tmp_start_epoch_time."</td>\n";
+						$content .= "	<td class='middle right no-wrap'>".$row['start_date_formatted']."</td>\n";
+						$content .= "	<td class='middle right no-wrap hide-md-dn'>".$row['start_time_formatted']."</td>\n";
 					}
 				//tta (time to answer)
 					if (permission_exists('xml_cdr_tta')) {
-						$content .= "	<td class='middle right hide-md-dn'>".(($row['tta'] > 0) ? $row['tta']."s" : "&nbsp;")."</td>\n";
+						$content .= "	<td class='middle right hide-md-dn'>".(($row['tta'] >= 0) ? $row['tta']."s" : "&nbsp;")."</td>\n";
 					}
 				//duration
 					if (permission_exists('xml_cdr_duration')) {
@@ -816,11 +812,7 @@
 					}
 
 					$content .= "</tr>\n";
-
 				//show the leg b only to those with the permission
-					if (!permission_exists('xml_cdr_lose_race') && $row['hangup_cause'] == 'LOSE_RACE') {
-						$content = '';
-					}
 					if ($row['leg'] == 'a') {
 						echo $content;
 					}
